+++ Changelog.md
@@ -13,3 +13,13 @@ http://oppnkay27.bkt.clouddn.com/20180517152648931917150.png
 3. others...
 
 http://oppnkay27.bkt.clouddn.com/20180517152657143787069.png
+
+=== 05/20 ===
+
+1. feat: Highlight acitve line;
+2. feat: Add the line numbers;
+3. feat: Add the placeholder;
+4. feat: Complete the following functions  undo/redo/bold/italic/line-through/block-quote/uppercase/capitalize/lowercase/h1-h5/ordered-list/unordered-list/divider/link/image(exclude: imglink)/inlinecode/timestamp/empty;
+5. other...
+
+http://oppnkay27.bkt.clouddn.com/20180520152680855740911.png